requirements.txt
accelerate==0.27.2
animeface==2.0.2
bitsandbytes==0.42.0
compel==2.0.2
controlnet_aux @ git+https://github.com/Tenpi/controlnet_aux
diffusers @ git+https://github.com/huggingface/diffusers
datasets==2.18.0
evaluate==0.4.1
Flask==2.3.2
Flask_SocketIO==5.3.4
numpy==1.26.4
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==1.3.5
peft==0.9.0
piexif==1.1.3
Pillow==10.2.0
prodigyopt==1.0
psd_tools==1.9.31
pysaucenao==1.6.2
python_engineio==4.5.1
pytorch_lightning==2.2.0.post0
pytoshop==1.2.1
PyWavelets==1.4.1
Requests==2.31.0
safetensors==0.4.2
scikit_learn==1.4.1.post1
segment_anything==1.0
Send2Trash==1.8.0
setuptools==69.1.1
scikit_image==0.22.0
tensorflow==2.15.0
tensorflow_macos==2.15.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
